Pushing a months work. Audiostream finshed and working well; added some dsp tools, fixed up the resonator system to support (nearly) arbitrary windows, some typos and warnings, enable&disable on comboboxes. Some opengl errors removed, completely reworked dsp windows with 19 different types and a parameterized interface, removed additions to std namespace (disallowed), added some exotic math routines, added some octave stuff.

+++ CSerializer.h
@@ -115,7 +115,8 @@
 	#include <memory>
 	#include <exception>
 	#include <stdexcept>
-	#include <cpl/stdext.h>
+	#include "stdext.h"
+	#include "PlatformSpecific.h"
 
 	namespace cpl
 	{
@@ -314,7 +315,7 @@
 
 		private:
 			const BinaryBuilder::byte * contents;
-			std::size_t contentSize;
+			std::int64_t contentSize;
 		};
 
 		class ISerializerSystem
